move functions to more appropriate destination
