+++ contracts/metastaking_contract.py
@@ -11,9 +11,8 @@ from erdpy.accounts import Account, Address
 from erdpy.contracts import SmartContract, CodeMetadata
 from erdpy.proxy import ElrondProxy
 from erdpy.transactions import Transaction
-from utils.utils_chain import print_transaction_hash, print_test_step_fail, print_test_step_pass, \
-    print_test_substep
-from utils.utils_chain import print_warning
+from utils.utils_chain import print_transaction_hash
+from utils.utils_generic import print_test_step_fail, print_test_step_pass, print_test_substep, print_warning
 
 
 class MetaStakingContract(DEXContractInterface):
