+++ contracts/proxy_deployer_contract.py
@@ -4,8 +4,8 @@ import traceback
 from arrows.stress.contracts.contract import load_code_as_hex
 from contracts.contract_identities import DEXContractInterface, RouterContractVersion
 from utils.utils_tx import prepare_contract_call_tx, send_contract_call_tx, get_deployed_address_from_event
-from utils.utils_chain import print_transaction_hash, print_test_step_fail, print_warning, \
-    print_test_step_pass
+from utils.utils_chain import print_transaction_hash
+from utils.utils_generic import print_test_step_fail, print_test_step_pass, print_warning
 from erdpy.accounts import Account, Address
 from erdpy.contracts import SmartContract, CodeMetadata
 from erdpy.proxy import ElrondProxy
