+++ contracts/simple_lock_energy_contract.py
@@ -5,8 +5,8 @@ from arrows.stress.contracts.contract import load_code_as_hex
 from contracts.contract_identities import DEXContractInterface
 from utils.utils_tx import prepare_contract_call_tx, send_contract_call_tx, \
     multi_esdt_endpoint_call, endpoint_call
-from utils.utils_chain import print_warning, print_transaction_hash, print_test_step_fail, \
-    print_test_step_pass, print_test_substep
+from utils.utils_chain import print_transaction_hash
+from utils.utils_generic import print_test_step_fail, print_test_step_pass, print_test_substep, print_warning
 from erdpy.accounts import Account, Address
 from erdpy.contracts import CodeMetadata, SmartContract
 from erdpy.proxy import ElrondProxy
